requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
frozenlist==1.3.3
idna==3.4
loguru==0.7.0
magic-filter==1.0.9
multidict==6.0.4
psycopg2==2.9.6
pycodestyle==2.10.0
python-dotenv==1.0.0
pytz==2023.3
win32-setctime==1.1.0
yarl==1.9.1
